Move testingSqlClient into server
